Crate sator_reward

Reward for viewers.

Creator initializes show

Creator sets prove user can participate in quiz

Quiz results are put into contract

After some lock time, it is possible to claim reward from each winner quiz

Derivation rules:

 fn find_program_address_for_pubkey(
     seed: &Pubkey,
     program_id: &ProgramPubkey,
 ) -> (ProgramDerivedPubkey, u8) {
     Pubkey::find_program_address(&[&seed.to_bytes()[..32]], program_id)
 }
 let show_authority = Pubkey::find_program_address_for_pubkey(&show.pubkey(), &program_id());
 let token_account_show =
     Pubkey::create_with_seed(&show_authority.0, "Show::token_account", &spl_token::id())?;
 let (show_authority_pubkey, _) = Pubkey::find_program_address_for_pubkey(show, &program_id());
         
 let (show_authority_pubkey, _) =
     Pubkey::find_program_address_for_pubkey(&show.pubkey(), &program_id());
 let (viewer_pubkey, _) = Pubkey::create_with_seed_for_pubkey(
     &show_authority_pubkey,
     &input.user_wallet,
     &program_id(),
 )?;
 fn create_with_seed_index(
     base: &Pubkey,
     seed: &str,
     seed_index: u64,
     owner: &ProgramPubkey,
 ) -> Result<(ProgramDerivedPubkey, String), PubkeyError> {
     let seed = format!("{}{:?}", seed, seed_index);
     let pubkey = Pubkey::create_with_seed(base, &seed, owner)?;
     Ok((pubkey, seed))
 }
 let (quiz_pubkey, _) = Pubkey::create_with_seed_index(
     &show_authority_pubkey,
     "Show::quizes",
     show_quizes_index as u64,
     &program_id(),
 )?;

Modules

errors

instruction

Program instruction state

state

Program owned state

types

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

id

Returns the program ID

program_id