Function sator_reward::instruction::initialize_quiz

pub fn initialize_quiz(
    owner: &SignerPubkey, 
    show: &Pubkey, 
    show_quizes_index: u16, 
    winners: Vec<Pubkey>, 
    input: InitializeQuizInput
) -> Result<Instruction, ProgramError>

Creates Instruction::InitializeQuiz instruction which initializes quiz with results. Validates winner is viewer. show’s quizes latest number must be provided. Winners and viewers must be in same corresponding order (zip should work), and less or equal to 5.

Instruction does not forces specified locked amount to be presented on on Show::token_account which is risk for user will not be payed. Accounts:

system_program - program, implicit to create accounts

sysvar_rent - program, implicit ensure that quiz are rent exempt.

clock - program, implicit to calculate prize won time

owner - signer, payer and owner of show.

show - used to validate owner and quiz and tak

show_authority - implicit program derived account from 32 bytes show public key based program_id.

quiz - mutable, derived from show + ’ShowState::index`