Function sator_reward::instruction::initialize_show

pub fn initialize_show(
    owner: &SignerPubkey, 
    show: &SignerPubkey, 
    mint: &MintPubkey, 
    input: InitializeShowInput
) -> Result<Instruction, ProgramError>

Creates Instruction::InitializeShow instruction which initializes show and shows’ token_account

Accounts:

system_program - program, implicit to create accounts

sysvar_rent - program, implicit ensure that token_account and show are rent exempt.

spl_token_program - program, implicit spl token program to initialize token_account.

owner - signer, payer and owner of show.

show - mutable, signer not initialized not created account for show data.

show_authority - implicit program derived account from 32 bytes show public key based program_id.

token_account - implicit, mutable, derived not created program derived account to create spl_token under show_authority.

mint - used to initialize token_account for reference